CustomerLogOnPage class deleted, updated SignInPage class
