+++ jdqz3-projekt-mk3/src/main/java/pages/SignInPage.java
@@ -11,16 +11,19 @@ public class SignInPage extends BasePage {
     private By signInPasswordSelector = By.id("signin_password");
     private By signInButtonSelector = By.id("genericLogin-button");
     private By loginErrorMessageSelector = By.id("loginError");
+    private By newCustomerButtonSelector = By.cssSelector("a.btn.btn-default.login-btn");
 
     private TextInput customerEmailAddress;
     private TextInput signInPassword;
     private Button signInButton;
     private Label loginErrorMessage;
+    private Button newCustomerButton;
 
     public SignInPage(WebDriver driver) {
         super(driver);
         this.customerEmailAddress = new TextInput(driver, customerEmailAddressSelector);
         this.signInPassword = new TextInput(driver, signInPasswordSelector);
+        this.newCustomerButton = new Button(driver, newCustomerButtonSelector);
     }
 
     public void clickSignInButton(){
@@ -32,5 +35,9 @@ public class SignInPage extends BasePage {
         this.loginErrorMessage = new Label(driver, loginErrorMessageSelector);
         String messageForEmptyCredentials = loginErrorMessage.readLabel();
         return messageForEmptyCredentials;
+    }
+
+    public void createNewAccount(){
+        this.newCustomerButton.click();
     }
 }
